perf: improve structure between LLMStreamService and BaseOpenAIService
